refactor: split template loaders
