+++ prompti/loader/__init__.py
@@ -1,0 +1,23 @@
+"""Template loaders package."""
+
+from .filesystem import FileSystemLoader
+from .memory import MemoryLoader
+from .http import HTTPLoader
+from .promptlayer import PromptLayerLoader
+from .langfuse import LangfuseLoader
+from .pezzo import PezzoLoader
+from .agenta import AgentaLoader
+from .github_repo import GitHubRepoLoader
+from .local_git_repo import LocalGitRepoLoader
+
+__all__ = [
+    "FileSystemLoader",
+    "MemoryLoader",
+    "HTTPLoader",
+    "PromptLayerLoader",
+    "LangfuseLoader",
+    "PezzoLoader",
+    "AgentaLoader",
+    "GitHubRepoLoader",
+    "LocalGitRepoLoader",
+]
